fixes on backend

+++ src/main/java/com/fdifrison/catan/dicecounter/controller/GameController.java
@@ -28,21 +28,21 @@ public class GameController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<GameDTO> getGameById(@PathVariable Long id) {
+    public ResponseEntity<GameDTO> getGameById(@PathVariable Integer id) {  // Changed from Long
         return gameService.getGameById(id)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
     }
 
     @GetMapping("/{id}/duration")
-    public ResponseEntity<Long> getGameDuration(@PathVariable Long id) {
+    public ResponseEntity<Long> getGameDuration(@PathVariable Integer id) {  // Changed from Long
         return gameService.getGameById(id)
                 .map(game -> ResponseEntity.ok(gameService.calculateGameDuration(game)))
                 .orElse(ResponseEntity.notFound().build());
     }
 
     @GetMapping("/{id}/slowest-player")
-    public ResponseEntity<String> getSlowestPlayer(@PathVariable Long id) {
+    public ResponseEntity<String> getSlowestPlayer(@PathVariable Integer id) {  // Changed from Long
         return ResponseEntity.ok(gameService.findSlowestPlayer(id));
     }
 
@@ -53,19 +53,19 @@ public class GameController {
     }
 
     @PostMapping("/{gameId}/turns")
-    public ResponseEntity<TurnDTO> recordTurn(@PathVariable Long gameId, @RequestBody TurnCreateDTO turnCreateDTO) {
+    public ResponseEntity<TurnDTO> recordTurn(@PathVariable Integer gameId, @RequestBody TurnCreateDTO turnCreateDTO) {  // Changed from Long
         TurnDTO recordedTurn = gameService.recordTurn(gameId, turnCreateDTO);
         return ResponseEntity.ok(recordedTurn);
     }
 
     @PutMapping("/{id}/end")
-    public ResponseEntity<GameDTO> endGame(@PathVariable Long id, @RequestBody EndGameDTO endGameDTO) {
+    public ResponseEntity<GameDTO> endGame(@PathVariable Integer id, @RequestBody EndGameDTO endGameDTO) {  // Changed from Long
         GameDTO endedGame = gameService.endGame(id, endGameDTO);
         return ResponseEntity.ok(endedGame);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteGame(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteGame(@PathVariable Integer id) {  // Changed from Long
         gameService.deleteGame(id);
         return ResponseEntity.noContent().build();
     }
