+++ src/main/java/com/fdifrison/catan/dicecounter/dto/TurnCreateDTO.java
@@ -3,7 +3,7 @@ package com.fdifrison.catan.dicecounter.dto;
 import java.time.Instant;
 
 public record TurnCreateDTO(
-        Long playerId,
+        Integer playerId,
         int turnNumber,
         Instant startTimestamp,
         Instant endTimestamp,
