+++ src/main/java/com/fdifrison/catan/dicecounter/dto/TurnDTO.java
@@ -3,8 +3,8 @@ package com.fdifrison.catan.dicecounter.dto;
 import java.time.Instant;
 
 public record TurnDTO(
-        Long id,
-        Long playerId,
+        Integer id,
+        Integer playerId,
         int turnNumber,
         Instant startTimestamp,
         Instant endTimestamp
